backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
stellar-sdk>=12.0.0,<14.0.0
sqlalchemy[asyncio]>=2.0.36,<3.0.0
asyncpg>=0.30.0,<1.0.0
python-dotenv>=1.0.1,<2.0.0
httpx>=0.28.0,<1.0.0
websockets>=14.0,<16.0.0
pydantic>=2.10.0,<3.0.0
alembic>=1.14.0,<2.0.0
python-socketio>=5.11.0,<6.0.0
tenacity>=8.2.0,<10.0.0
cachetools>=5.3.0,<6.0.0
redis>=5.0.0,<6.0.0
